requirements.txt
wikipedia
gradio
transformers
sentence-transformers
faiss-cpu
